chore: 파일 및 폴더 물리구조 변경
